changed names for better clarity
used broadcasting for cleaner code
